@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-	<TargetFrameworks>net6.0;net8.0</TargetFrameworks>
+	<TargetFrameworks>net8.0;net9.0</TargetFrameworks>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <Title>$(AssemblyName)</Title>
@@ -16,10 +16,9 @@
     <PackageTags>ModdableWebServer;webserver;web</PackageTags>
     <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
     <PackageReleaseNotes>$(Version)!</PackageReleaseNotes>
-    <Version>1.8</Version>
+    <Version>2.0</Version>
     <FileVersion>$(Version)</FileVersion>
     <AssemblyVersion>$(Version)</AssemblyVersion>
-	<LangVersion>12</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
@@ -33,14 +32,14 @@
     </None>
   </ItemGroup>
 
-  <ItemGroup Condition="'$(TargetFramework)' == 'net6.0'">
-    <PackageReference Include="L" Version="5.0.0" />
-    <PackageReference Include="NetCoreServer" Version="6.7.0" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="NetCoreServer" Version="8.0.7" />
+		<PackageReference Include="Serilog" Version="4.3.0" />
+	</ItemGroup>
 
-<ItemGroup Condition="'$(TargetFramework)' == 'net8.0'">
-	<PackageReference Include="L" Version="5.0.0" />
-	<PackageReference Include="NetCoreServer" Version="8.0.7" />
-</ItemGroup>
+	<ItemGroup>
+	  <Using Include="NetCoreServer" />
+	  <Using Include="Serilog" />
+	</ItemGroup>
 
 </Project>
