@@ -1,141 +1,46 @@
 ﻿using ModdableWebServer.Attributes;
 using ModdableWebServer.Helper;
-using NetCoreServer;
+using ModdableWebServer.Interfaces;
+using ModdableWebServer.Sessions;
 using System.Net;
 using System.Net.Sockets;
 using System.Reflection;
 
-namespace ModdableWebServer.Servers
+namespace ModdableWebServer.Servers;
+
+public class HTTP_Server : HttpServer, IServer
 {
-    public class HTTP_Server : HttpServer
+    public HTTP_Server(string address, int port) : base(address, port) { }
+    public HTTP_Server(IPAddress address, int port) : base(address, port) { }
+    public HTTP_Server(DnsEndPoint endpoint) : base(endpoint) { }
+    public HTTP_Server(IPEndPoint endPoint) : base(endPoint) { }
+    public Dictionary<HTTPAttribute, MethodInfo> HTTPAttributeToMethods { get; } = [];
+    public Dictionary<HTTPHeaderAttribute, MethodInfo> HeaderAttributeToMethods { get; } = [];
+
+    public bool DoReturn404IfFail { get; set; } = true;
+
+    #region Attribute Controls
+
+    public void OverrideAttributes(Assembly assembly)
     {
-        #region Events
-        public EventHandler<(HttpRequest request, string error)>? ReceivedRequestError;
-        public EventHandler<SocketError>? OnSocketError;
-        public EventHandler<HttpRequest>? ReceivedFailed;
-        public event EventHandler<(string address, int port)>? Started;
-        public event EventHandler? Stopped;
-        #endregion
-        public Dictionary<HTTPAttribute, MethodInfo> AttributeToMethods = [];
-        public Dictionary<HTTPHeaderAttribute, MethodInfo> HeaderAttributeToMethods = [];
-
-        public bool DoReturn404IfFail = true;
-        public HTTP_Server(string address, int port) : base(address, port)
-            => Load();
-
-        public HTTP_Server(IPAddress address, int port) : base(address, port)
-            => Load();
-
-        public HTTP_Server(DnsEndPoint endpoint) : base(endpoint)
-            => Load();
-
-        public HTTP_Server(IPEndPoint endPoint) : base(endPoint)
-            => Load();
-
-        #region Attribute Controls
-        private void Load()
-        {
-            Assembly? asm = Assembly.GetAssembly(typeof(HTTPAttribute));
-            AttributeToMethods = AttributeMethodHelper.GetMethodAndAttribute<HTTPAttribute>(asm);
-            HeaderAttributeToMethods = AttributeMethodHelper.GetMethodAndAttribute<HTTPHeaderAttribute>(asm);
-        }
-
-        public void OverrideAttribute(Assembly assembly)
-        {
-            AttributeToMethods.Override(assembly);
-        }
-
-        public void MergeAttribute(Assembly assembly)
-        {
-            AttributeToMethods.Merge(assembly);
-        }
-
-        public void ClearAttribute()
-        {
-            AttributeToMethods.Clear();
-        }
-        public void OverrideHeaderAttribute(Assembly assembly)
-        {
-            HeaderAttributeToMethods.Override(assembly);
-        }
-
-        public void MergeHeaderAttribute(Assembly assembly)
-        {
-            HeaderAttributeToMethods.Merge(assembly);
-        }
-
-        public void ClearHeaderAttribute()
-        {
-            HeaderAttributeToMethods.Clear();
-        }
-
-        public void OverrideAttributes(Assembly assembly)
-        {
-            HeaderAttributeToMethods.Override(assembly);
-            AttributeToMethods.Override(assembly);
-        }
-
-        public void MergeAttributes(Assembly assembly)
-        {
-            HeaderAttributeToMethods.Merge(assembly);
-            AttributeToMethods.Merge(assembly);
-        }
-
-        public void ClearAttributes()
-        {
-            HeaderAttributeToMethods.Clear();
-            AttributeToMethods.Clear();
-        }
-
-        #endregion
-        #region Overrides
-        protected override void OnStarted() => Started?.Invoke(this, (Address, Port));
-
-        protected override void OnStopped() => Stopped?.Invoke(this, new());
-
-        protected override TcpSession CreateSession() { return new Session(this); }
-
-        protected override void OnError(SocketError error) => OnSocketError?.Invoke(this, error);
-        #endregion
-
-        public class Session(HttpServer server) : HttpSession(server)
-        {
-            public HTTP_Server HTTP_Server => (HTTP_Server)this.Server;
-
-            #region Overrides
-            protected override void OnReceivedRequest(HttpRequest request)
-            {
-                if (request.Method == "GET" && !request.Url.Contains('?') && this.Cache.FindPath(request.Url))
-                {
-                    var cache = this.Cache.Find(request.Url);
-                    // Check again to make sure.
-                    if (cache.Item1)
-                        this.SendAsync(cache.Item2);
-                }
-
-                ServerStruct serverStruct = new()
-                {
-                    HTTP_Session = this,
-                    Response = this.Response,
-                    Enum = ServerEnum.HTTP
-                };
-
-                bool IsSent = serverStruct.SendRequestHTTP(request, HTTP_Server.AttributeToMethods);
-                bool IsSent_header = serverStruct.SendRequestHTTPHeader(request, HTTP_Server.HeaderAttributeToMethods);
-
-                DebugPrinter.Debug("[HttpSession.OnReceivedRequest] Request sent!");
-
-                if (!IsSent && !IsSent_header)
-                    HTTP_Server.ReceivedFailed?.Invoke(this, request);
-
-                if (HTTP_Server.DoReturn404IfFail && (!IsSent && !IsSent_header))
-                    SendResponse(Response.MakeErrorResponse(404));
-            }
-
-            protected override void OnReceivedRequestError(HttpRequest request, string error) => HTTP_Server.ReceivedRequestError?.Invoke(this, (request, error));
-
-            protected override void OnError(SocketError error) => HTTP_Server.OnSocketError?.Invoke(null, error);
-            #endregion
-        }
+        HeaderAttributeToMethods.Override(assembly);
+        HTTPAttributeToMethods.Override(assembly);
     }
+
+    public void MergeAttributes(Assembly assembly)
+    {
+        HeaderAttributeToMethods.Merge(assembly);
+        HTTPAttributeToMethods.Merge(assembly);
+    }
+
+    public void ClearAttributes()
+    {
+        HeaderAttributeToMethods.Clear();
+        HTTPAttributeToMethods.Clear();
+    }
+    #endregion
+    protected override void OnStarted() => ServerEvents.OnStarted(this);
+    protected override void OnStopped() => ServerEvents.OnStopped(this);
+    protected override TcpSession CreateSession() => new HTTP_Session(this);
+    protected override void OnError(SocketError error) => ServerEvents.OnSocketError(this, error);
 }
