@@ -1,15 +1,14 @@
-﻿using ModdableWebServer.Attributes;
-using System.Reflection;
+﻿using System.Reflection;
 
 namespace ModdableWebServer.Helper;
 
 public static class AttributeMethodHelper
 {
-    public static void Merge(this Dictionary<HTTPAttribute, MethodInfo> keyValues, Assembly? assembly)
+    public static void Merge<T>(this Dictionary<T, MethodInfo> keyValues, Assembly? assembly) where T : Attribute
     {
         ArgumentNullException.ThrowIfNull(assembly);
 
-        var attrib_methods = GetMethodAndAttribute<HTTPAttribute>(assembly);
+        var attrib_methods = GetMethodAndAttribute<T>(assembly);
         foreach (var (attrib, method) in attrib_methods)
         {
             if (attrib == null)
@@ -17,54 +16,17 @@ public static class AttributeMethodHelper
             if (method == null)
                 continue;
             if (!keyValues.TryAdd(attrib, method))
-                DebugPrinter.Debug($"[Merge HTTP] Cannot merge {attrib.method} ,{attrib.url}.");
+                Log.Warning($"Cannot merge {attrib}.");
             else
-                DebugPrinter.Debug($"[Merge HTTP] {attrib.method}, {attrib.url} = {method}");
+                Log.Verbose($"{attrib} = {method}");
         }
     }
 
-    public static void Merge(this Dictionary<HTTPHeaderAttribute, MethodInfo> keyValues, Assembly? assembly)
+    public static void Override<T>(this Dictionary<T, MethodInfo> keyValues, Assembly? assembly) where T : Attribute
     {
         ArgumentNullException.ThrowIfNull(assembly);
 
-        var attrib_methods = GetMethodAndAttribute<HTTPHeaderAttribute>(assembly);
-        foreach (var (attrib, method) in attrib_methods)
-        {
-            if (attrib == null)
-                continue;
-            if (method == null)
-                continue;
-            if (!keyValues.TryAdd(attrib, method))
-                DebugPrinter.Debug($"[Merge HTTPHeader] Cannot merge {attrib.method} ,{attrib.url}.");
-            else
-                DebugPrinter.Debug($"[Merge HTTPHeader] {attrib.method}, {attrib.url} = {method}");
-        }
-    }
-
-
-    public static void Merge(this Dictionary<string, MethodInfo> keyValues, Assembly? assembly)
-    {
-        ArgumentNullException.ThrowIfNull(assembly);
-
-        var attrib_methods = GetMethodAndAttribute<WSAttribute>(assembly);
-        foreach (var (attrib, method) in attrib_methods)
-        {
-            if (attrib == null)
-                continue;
-            if (method == null)
-                continue;
-            if (!keyValues.TryAdd(attrib.url, method))
-                DebugPrinter.Debug($"[Merge WS] Cannot merge {attrib.url}.");
-            else
-                DebugPrinter.Debug($"[Merge WS] {attrib.url} = {method}");
-        }
-    }
-
-    public static void Override(this Dictionary<HTTPAttribute, MethodInfo> keyValues, Assembly? assembly)
-    {
-        ArgumentNullException.ThrowIfNull(assembly);
-
-        var attrib_methods = GetMethodAndAttribute<HTTPAttribute>(assembly);
+        var attrib_methods = GetMethodAndAttribute<T>(assembly);
         foreach (var (attrib, method) in attrib_methods)
         {
             if (attrib == null)
@@ -75,45 +37,7 @@ public static class AttributeMethodHelper
             if (!keyValues.TryAdd(attrib, method))
                 throw new Exception("KeyValues adding was not possible (Even after removed old data)");
             else
-                DebugPrinter.Debug($"[Override HTTP] {attrib.method}, {attrib.url} = {method}");
-        }
-    }
-
-    public static void Override(this Dictionary<HTTPHeaderAttribute, MethodInfo> keyValues, Assembly? assembly)
-    {
-        ArgumentNullException.ThrowIfNull(assembly);
-
-        var attrib_methods = GetMethodAndAttribute<HTTPHeaderAttribute>(assembly);
-        foreach (var (attrib, method) in attrib_methods)
-        {
-            if (attrib == null)
-                continue;
-            if (method == null)
-                continue;
-            keyValues.Remove(attrib);
-            if (!keyValues.TryAdd(attrib, method))
-                throw new Exception("KeyValues adding was not possible (Even after removed old data)");
-            else
-                DebugPrinter.Debug($"[Override HTTPHeader] {attrib.method}, {attrib.url} = {method}");
-        }
-    }
-
-    public static void Override(this Dictionary<string, MethodInfo> keyValues, Assembly? assembly)
-    {
-        ArgumentNullException.ThrowIfNull(assembly);
-
-        var attrib_methods = GetMethodAndAttribute<WSAttribute>(assembly);
-        foreach (var (attrib, method) in attrib_methods)
-        {
-            if (attrib == null)
-                continue;
-            if (method == null)
-                continue;
-            keyValues.Remove(attrib.url);
-            if (!keyValues.TryAdd(attrib.url, method))
-                throw new Exception("KeyValues adding was not possible (Even after removed old data)");
-            else
-                DebugPrinter.Debug($"[Override WS] {attrib.url} = {method}");
+                Log.Verbose($"[Override] {attrib} = {method}");
         }
     }
 
@@ -123,10 +47,13 @@ public static class AttributeMethodHelper
 
         Dictionary<T, MethodInfo> values = [];
 
-        var methods = assembly.GetTypes().SelectMany(x => x.GetMethods()).ToArray();
-        //hacky  method to get custom attrib that is either void or bool (bool represent HTTP, void represent WS)
-        var basemethods = methods.Where(x => x.IsDefined(typeof(T)) && (x.ReturnType == typeof(void) || x.ReturnType == typeof(bool))).ToArray();
-        foreach (var method in basemethods)
+        foreach (var method in assembly.GetTypes().SelectMany(select =>
+            select.GetMethods().Where(method => 
+                method.IsDefined(typeof(T)) && 
+                (method.ReturnType == typeof(void) || method.ReturnType == typeof(bool)
+                )
+            )
+        ))
         {
             if (method == null)
                 continue;
