@@ -1,34 +1,37 @@
 ﻿#if DEBUG
 using ModdableWebServer.Attributes;
 using ModdableWebServer.Helper;
-using NetCoreServer;
+using ModdableWebServer.Senders;
 
 namespace ModdableWebServer;
 
 public class TEST
 {
 
-    [WS("/ws/{test}")]
-    public static void Ws(WebSocketStruct ws_Struct)
+    [WS("/ws/{test}",  WebSocketMethodListen.All)]
+    public static void Ws(WebSocketSender sender)
     {
         Console.WriteLine("Headers:");
-        foreach (var item in ws_Struct.Request.Headers)
+        foreach (var item in sender.Headers)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
         Console.WriteLine("Parameters:");
-
-        foreach (var item in ws_Struct.Request.Parameters)
+        foreach (var item in sender.Parameters)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
-        Console.WriteLine("IsConnected: " + ws_Struct.IsConnected);
-        Console.WriteLine("IsConnecting: " + ws_Struct.IsConnecting);
-        Console.WriteLine("IsClosed: " + ws_Struct.IsClosed);
-        Console.WriteLine(ws_Struct.Request.Url);
-        Console.WriteLine(ws_Struct.WSRequest?.buffer);
-        Console.WriteLine(ws_Struct.WSRequest?.offset);
-        Console.WriteLine(ws_Struct.WSRequest?.size);
+        Console.WriteLine("Current Listening: " + sender.CurrentMethod);
+        if (sender.CurrentMethod.HasFlag(WebSocketMethodListen.Close))
+        {
+            Console.WriteLine("CloseStatus: " + sender.CloseStatus);
+        }
+        if (sender.CurrentMethod > WebSocketMethodListen.Received)
+        {
+            Console.WriteLine(sender.Buffer.Length);
+            Console.WriteLine(sender.Offset);
+            Console.WriteLine(sender.Size);
+        }
     }
 
 
@@ -36,38 +39,58 @@ public class TEST
     [HTTP("GET", "/test")]
     [HTTP("GET", "/test1")]
     [HTTP("GET", "/test11")]
-    public static bool TEST1(HttpRequest _, ServerStruct serverStruct)
+    public static bool TEST1(ServerSender sender)
     {
         Console.WriteLine("Headers:");
-        foreach (var item in serverStruct.Headers)
+        foreach (var item in sender.Headers)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
         Console.WriteLine("Parameters:");
-
-        foreach (var item in serverStruct.Parameters)
+        foreach (var item in sender.Parameters)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
 
         Console.WriteLine("TEST");
-        serverStruct.Response.MakeOkResponse();
-        ResponseSender.SendResponse(serverStruct);
+        sender.Response.MakeOkResponse();
+        sender.SendResponse();
+        return true;
+    }
+
+
+
+    [HTTP("GET", "/{!args}")]
+    public static bool Lol(ServerSender sender)
+    {
+        Console.WriteLine("Headers:");
+        foreach (var item in sender.Headers)
+        {
+            Console.WriteLine(item.Key + " = " + item.Value);
+        }
+        Console.WriteLine("Parameters:");
+        foreach (var item in sender.Parameters)
+        {
+            Console.WriteLine(item.Key + " = " + item.Value);
+        }
+
+        Console.WriteLine("!ARGS");
+        sender.Response.MakeOkResponse();
+        sender.SendResponse();
         return true;
     }
 
 
     [HTTP("DELETE", "/del")]
-    public static bool DEL(HttpRequest _, ServerStruct serverStruct)
+    public static bool DEL(ServerSender sender)
     {
         Console.WriteLine("Headers:");
-        foreach (var item in serverStruct.Headers)
+        foreach (var item in sender.Headers)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
         Console.WriteLine("Parameters:");
-
-        foreach (var item in serverStruct.Parameters)
+        foreach (var item in sender.Parameters)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
@@ -75,23 +98,21 @@ public class TEST
         Console.WriteLine("TEST");
         ResponseCreator response = new(200);
         response.SetBody("balls");
-        serverStruct.Response = response.GetResponse();
-        serverStruct.SendResponse();
+        sender.SendResponse(response);
         return true;
     }
 
     [HTTPHeader("GET", "/testheader2", "test")]
     [HTTPHeader("GET", "/testheader", "test")]
-    public static bool TEST_HEADER(HttpRequest _, ServerStruct serverStruct)
+    public static bool TEST_HEADER(ServerSender sender)
     {
         Console.WriteLine("Headers:");
-        foreach (var item in serverStruct.Headers)
+        foreach (var item in sender.Headers)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
         Console.WriteLine("Parameters:");
-
-        foreach (var item in serverStruct.Parameters)
+        foreach (var item in sender.Parameters)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
@@ -99,23 +120,22 @@ public class TEST
         Console.WriteLine("TEST HEADER");
         ResponseCreator response = new();
         response.SetBody("test header");
-        serverStruct.Response = response.GetResponse();
-        ResponseSender.SendResponse(serverStruct);
+        sender.SendResponse(response);
         return true;
     }
 
     // IF we would use the same url we would have been down, since the other one is triggered before this one.
-    [HTTPHeader("GET", "/testheader2", "test","myvalue")]
-    public static bool TEST_HEADER2(HttpRequest _, ServerStruct serverStruct)
+    [HTTPHeader("GET", "/testheader2", "test", "myvalue")]
+    [HTTPHeader("GET", "/testheader2", "test2", "jtw.", true)]
+    public static bool TEST_HEADER2(ServerSender sender)
     {
         Console.WriteLine("Headers:");
-        foreach (var item in serverStruct.Headers)
+        foreach (var item in sender.Headers)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
         Console.WriteLine("Parameters:");
-
-        foreach (var item in serverStruct.Parameters)
+        foreach (var item in sender.Parameters)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
@@ -123,23 +143,21 @@ public class TEST
         Console.WriteLine("TEST HEADER 2");
         ResponseCreator response = new();
         response.SetBody("test header 2");
-        serverStruct.Response = response.GetResponse();
-        ResponseSender.SendResponse(serverStruct);
+        sender.SendResponse(response);
         return true;
     }
 
 
     [HTTP("GET", "/test2/{test}")]
-    public static bool Test2(HttpRequest _, ServerStruct serverStruct)
+    public static bool Test2(ServerSender sender)
     {
         Console.WriteLine("Headers:");
-        foreach (var item in serverStruct.Headers)
+        foreach (var item in sender.Headers)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
         Console.WriteLine("Parameters:");
-
-        foreach (var item in serverStruct.Parameters)
+        foreach (var item in sender.Parameters)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
@@ -151,23 +169,21 @@ public class TEST
             { "key", "Value" }
         });
         response.SetBody("this is a proper test, and making things");
-        serverStruct.Response = response.GetResponse();
-        ResponseSender.SendResponse(serverStruct);
+        sender.SendResponse(response);
         return true;
     }
 
 
     [HTTP("GET", "/test3/{test}/test?aa=yes&asd={xx}")]
-    public static bool Test3(HttpRequest _, ServerStruct serverStruct)
+    public static bool Test3(ServerSender sender)
     {
         Console.WriteLine("Headers:");
-        foreach (var item in serverStruct.Headers)
+        foreach (var item in sender.Headers)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
         Console.WriteLine("Parameters:");
-
-        foreach (var item in serverStruct.Parameters)
+        foreach (var item in sender.Parameters)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
@@ -179,22 +195,20 @@ public class TEST
             { "key", "Value" }
         });
         response.SetBody("this is a proper test, and making things");
-        serverStruct.Response = response.GetResponse();
-        ResponseSender.SendResponse(serverStruct);
+        sender.SendResponse(response);
         return true;
     }
 
-    [HTTP("GET", "/test3/{test}/test2?{args}")]
-    public static bool Test3args(HttpRequest _, ServerStruct serverStruct)
+    [HTTP("GET", "/test3/{test}/test2?{!args}")]
+    public static bool Test3args(ServerSender sender)
     {
         Console.WriteLine("Headers:");
-        foreach (var item in serverStruct.Headers)
+        foreach (var item in sender.Headers)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
         Console.WriteLine("Parameters:");
-
-        foreach (var item in serverStruct.Parameters)
+        foreach (var item in sender.Parameters)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
@@ -206,23 +220,21 @@ public class TEST
             { "key", "Value" }
         });
         response.SetBody("this is a proper test, and making things");
-        serverStruct.Response = response.GetResponse();
-        ResponseSender.SendResponse(serverStruct);
+        sender.SendResponse(response);
         return true;
     }
 
 
-    [HTTP("GET", "/test3/{test}?{args}")]
-    public static bool Test3fullargs(HttpRequest _, ServerStruct serverStruct)
+    [HTTP("GET", "/test3/{test}?{!args}")]
+    public static bool Test3fullargs(ServerSender sender)
     {
         Console.WriteLine("Headers:");
-        foreach (var item in serverStruct.Headers)
+        foreach (var item in sender.Headers)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
         Console.WriteLine("Parameters:");
-
-        foreach (var item in serverStruct.Parameters)
+        foreach (var item in sender.Parameters)
         {
             Console.WriteLine(item.Key + " = " + item.Value);
         }
@@ -234,8 +246,7 @@ public class TEST
             { "key", "Value" }
         });
         response.SetBody("this is a proper test, and making things");
-        serverStruct.Response = response.GetResponse();
-        ResponseSender.SendResponse(serverStruct);
+        sender.SendResponse(response);
         return true;
     }
 
